Adding the remaining model classes
