+++ backend/orders/src/bootstrap.kt
@@ -1,6 +1,6 @@
 package challenge
 
-import challenge.model.Product
+import challenge.model.*
 import java.lang.Exception
 import java.util.*
 
@@ -14,7 +14,7 @@ class Order(val customer: Customer, val address: Address) {
         get() = items.sumByDouble { it.total }
 
     fun addProduct(product: Product, quantity: Int) {
-        var productAlreadyAdded = items.any { it.product == product }
+        val productAlreadyAdded = items.any { it.product == product }
         if (productAlreadyAdded)
             throw Exception("The product have already been added. Change the amount if you want more.")
 
@@ -38,50 +38,3 @@ class Order(val customer: Customer, val address: Address) {
     }
 }
 
-data class OrderItem(val product: Product, val quantity: Int) {
-    val total get() = product.price * quantity
-}
-
-data class Payment(val order: Order, val paymentMethod: PaymentMethod) {
-    val paidAt = Date()
-    val authorizationNumber = paidAt.time
-    val amount = order.totalAmount
-    val invoice = Invoice(order)
-}
-
-data class CreditCard(val number: String) : PaymentMethod
-
-interface PaymentMethod
-
-data class Invoice(val order: Order) {
-    val billingAddress: Address = order.address
-    val shippingAddress: Address = order.address
-}
-
-enum class ProductType {
-    PHYSICAL,
-    BOOK,
-    DIGITAL,
-    MEMBERSHIP
-}
-
-class Address
-
-class Customer
-
-//fun main(args : Array<String>) {
-//    val shirt = Product("Flowered t-shirt", ProductType.PHYSICAL, 35.00)
-//    val netflix = Product("Familiar plan", ProductType.MEMBERSHIP, 29.90)
-//    val book = Product("The Hitchhiker's Guide to the Galaxy", ProductType.BOOK, 120.00)
-//    val music = Product("Stairway to Heaven", ProductType.DIGITAL, 5.00)
-//
-//    val order = Order(Customer(), Address())
-//
-//    order.addProduct(shirt, 2)
-//    order.addProduct(netflix, 1)
-//    order.addProduct(book, 1)
-//    order.addProduct(music, 1)
-//
-//    order.pay(CreditCard("43567890-987654367"))
-//    // now, how to deal with shipping rules then?
-//}
