+++ backend/orders/src/model/Address.kt
@@ -1,0 +1,3 @@
+package challenge.model
+
+class Address
