+++ backend/orders/src/model/CreditCard.kt
@@ -1,0 +1,3 @@
+package challenge.model
+
+data class CreditCard(val number: String) : PaymentMethod
