+++ backend/orders/src/model/Customer.kt
@@ -1,0 +1,3 @@
+package challenge.model
+
+class Customer
