+++ backend/orders/src/model/Invoice.kt
@@ -1,0 +1,8 @@
+package challenge.model
+
+import challenge.Order
+
+data class Invoice(val order: Order) {
+    val billingAddress: Address = order.address
+    val shippingAddress: Address = order.address
+}
