+++ backend/orders/src/model/OrderItem.kt
@@ -1,0 +1,5 @@
+package challenge.model
+
+data class OrderItem(val product: Product, val quantity: Int) {
+    val total get() = product.price * quantity
+}
