+++ backend/orders/src/model/Payment.kt
@@ -1,0 +1,11 @@
+package challenge.model
+
+import challenge.Order
+import java.util.*
+
+data class Payment(val order: Order, val paymentMethod: PaymentMethod) {
+    private val paidAt = Date()
+    val authorizationNumber = paidAt.time
+    val amount = order.totalAmount
+    val invoice = Invoice(order)
+}
