+++ backend/orders/src/model/PaymentMethod.kt
@@ -1,0 +1,3 @@
+package challenge.model
+
+interface PaymentMethod
