+++ backend/orders/src/model/Product.kt
@@ -1,6 +1,4 @@
 package challenge.model
 
-import challenge.ProductType
-
 
 data class Product(val name: String, val type: ProductType, val price: Double)
