+++ backend/orders/src/model/ProductType.kt
@@ -1,0 +1,8 @@
+package challenge.model
+
+enum class ProductType {
+    PHYSICAL,
+    BOOK,
+    DIGITAL,
+    MEMBERSHIP
+}
